Fix minimizer renaming functions
When running the code through a minimizer (e g through Webpack/Terser),
the minimizer renames functions such as "$asBoolean". However, when the
function is called, it's hard coded as the string "$asBoolean", so the
generated code fails with "$asBoolean is not defined".

Make sure that the original $asBoolean function is included, and not
a renamed version of it, by making $asBoolean and friends string constants.

Fixes: https://github.com/fastify/fast-json-stringify/issues/338

Signed-off-by: David Henningsson <coding@diwic.se>